+++ tempreader.h
@@ -4,6 +4,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #define STATUS_OK 0
 #define STATUS_READING_ERROR 1
@@ -12,6 +13,6 @@
  * Obtem a leitura da CPU do Raspberry Pi
  * @return status da operação
  */
-int get_cpu_temperature(float* reading);
+uint8_t get_cpu_temperature(float* reading);
 
 #endif
